feat: 関数群を外だし
