+++ src/hooks/useFormText/useFormText.ts
@@ -1,14 +1,13 @@
 import { ChangeEvent, FormEventHandler, useState } from 'react';
-import { z } from 'zod';
 
 import {
-  Mode,
   FieldValue,
   FieldValueKey,
   FieldValueValidate,
   FieldState,
   UseFormTextArgs,
 } from '@/types/form';
+import { formValidator } from '@/utils/form';
 
 type UseFormText<T> = {
   fieldValue: T;
@@ -29,34 +28,8 @@ export const useFormText = <T extends FieldValue>({
 
   const [fieldState, setFieldState] = useState<FieldState<T>>({
     errors: undefined,
-    isValid: true,
+    isValid: false,
   });
-
-  const validator = (
-    mode: Mode,
-    validate: FieldValueValidate | undefined,
-    input: string
-  ) => {
-    if (typeof validate === 'undefined') {
-      return undefined;
-    }
-
-    switch (mode) {
-      case 'onChange': {
-        if (
-          typeof validate === 'undefined' ||
-          typeof validate.maxLength === 'undefined'
-        )
-          return;
-        try {
-          const maxLength = z.string().max(validate.maxLength.value);
-          maxLength.parse(input);
-        } catch (err: any) {
-          return validate.maxLength.message ?? 'invalid input';
-        }
-      }
-    }
-  };
 
   const onChange = (
     key: FieldValueKey<T>,
@@ -64,7 +37,7 @@ export const useFormText = <T extends FieldValue>({
     validate?: FieldValueValidate
   ) => {
     const input = evt.currentTarget.value;
-    const result = validator(mode, validate, input);
+    const result = formValidator(validate, input);
 
     if (typeof result !== 'undefined') {
       setFieldState({
